feat: 문자열 나누기 기능 개발
- 문자열 나누기 기능 개발
- 문자열 나누기 테스트 코드 개발
