+++ src/main/java/calculator/CalculatorController.java
@@ -1,0 +1,10 @@
+package calculator;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class CalculatorController {
+
+}
