--- src/main/java/calculator/StringCalculator.java
@@ -1,8 +1,0 @@
-package calculator;
-
-public class StringCalculator {
-
-    public Operand calculator(Operator operator, Operand left, Operand right) {
-        return operator.calculate(left, right);
-    }
-}
